app/requirements.txt
django
pytest
graphene-django
djangorestframework-jwt
djangorestframework
psycopg2-binary
gunicorn
whitenoise
tqdm